Extensive refactoring, for what it's worth
