+++ Cargo.toml
@@ -29,7 +29,7 @@ serde_json = "1.0.114"
 isatty = "0.1.8"
 toml = "0.4.5"
 serde_derive = "1.0.197"
-serde = "1.0.197"
+serde = { version = "1.0.197", features = ["derive"] }
 shlex = "0.1.1"
 dirs = "1.0"
 regex = "1.10.3"
